SSDCNet/requirements.txt
torch
numpy==1.14.0
pillow==6.2.0
scipy==1.0.0
pandas==0.22.0
torchvision
